Update 정수 제곱근 판별.cpp

+++ 프로그래머스/1/12934. 정수 제곱근 판별/정수 제곱근 판별.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 long long solution(long long n) {
-    long long sqrt_n = (long long)sqrt(n);
+    int sqrt_n = sqrt(n);
     
     if (pow((sqrt_n), 2) == n) {
         return pow(sqrt_n + 1, 2);
